+++ hoomd/hpmc/module_union_sphere.cc
@@ -13,13 +13,13 @@
 #include "ExternalFieldHarmonic.h"
 #include "ExternalFieldWall.h"
 
-#include "UpdaterClusters.h"
+#include "UpdaterGCA.h"
 #include "UpdaterMuVT.h"
 
 #ifdef ENABLE_HIP
 #include "ComputeFreeVolumeGPU.h"
 #include "IntegratorHPMCMonoGPU.h"
-#include "UpdaterClustersGPU.h"
+#include "UpdaterGCAGPU.h"
 #endif
 
 namespace hoomd
@@ -35,7 +35,7 @@ void export_union_sphere(pybind11::module& m)
     export_ComputeFreeVolume<ShapeUnion<ShapeSphere>>(m, "ComputeFreeVolumeSphereUnion");
     export_ComputeSDF<ShapeUnion<ShapeSphere>>(m, "ComputeSDFSphereUnion");
     export_UpdaterMuVT<ShapeUnion<ShapeSphere>>(m, "UpdaterMuVTSphereUnion");
-    export_UpdaterClusters<ShapeUnion<ShapeSphere>>(m, "UpdaterClustersSphereUnion");
+    export_UpdaterGCA<ShapeUnion<ShapeSphere>>(m, "UpdaterGCASphereUnion");
 
     export_ExternalFieldInterface<ShapeUnion<ShapeSphere>>(m, "ExternalFieldSphereUnion");
     export_HarmonicField<ShapeUnion<ShapeSphere>>(m, "ExternalFieldHarmonicSphereUnion");
@@ -45,7 +45,7 @@ void export_union_sphere(pybind11::module& m)
 
     export_IntegratorHPMCMonoGPU<ShapeUnion<ShapeSphere>>(m, "IntegratorHPMCMonoSphereUnionGPU");
     export_ComputeFreeVolumeGPU<ShapeUnion<ShapeSphere>>(m, "ComputeFreeVolumeSphereUnionGPU");
-    export_UpdaterClustersGPU<ShapeUnion<ShapeSphere>>(m, "UpdaterClustersSphereUnionGPU");
+    export_UpdaterGCAGPU<ShapeUnion<ShapeSphere>>(m, "UpdaterGCASphereUnionGPU");
 
 #endif
     }